docs: Home 프로브 측정 기준 정리 - #308
